send cat food all the time

+++ common/camera.py
@@ -7,9 +7,10 @@ import os
 def take_photo():
     capture = path.to('data/capture.jpeg')
 
-    os.remove(capture)
+    if os.path.isfile(capture):
+        os.remove(capture)
+    
     result = os.system('fswebcam -q -D 2 -r 1280x720 --jpeg 80 "{}"'.format(capture))
-
     if result == 0:
         return capture
     else:
